modularizing code, so far for config and updater. fundamental changes to how config is handled. making way for flask interface/daemon idea
